% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportSections.R
\name{createReportSections}
\alias{createReportSections}
\title{Script to Create Sections of the Report}
\usage{
createReportSections(
  connectionDetails,
  connectionDetailsMerge,
  cdmDatabaseSchema,
  databaseName = "",
  databaseId = "",
  outputFolder = "output",
  verboseMode = TRUE,
  accountUrl = "",
  accountKey = "",
  containerName = "",
  depth = 3
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{connectionDetailsMerge}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{databaseName}{String name of the database name. If blank, CDM_SOURCE table will be queried to try to obtain this.}

\item{databaseId}{ID of your database, this will be used as subfolder for the results.}

\item{outputFolder}{Path to store logs and SQL files}

\item{verboseMode}{Boolean to determine if the console will show all execution steps. Default = TRUE}

\item{accountUrl}{Web location of Azure storage account}

\item{accountKey}{SAS key for the Azure storage account}

\item{depth}{Depth of the directory structure in the Azure container (should be 3 per convention, but can vary in practice)}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want all temporary tables to be managed. Requires create/insert permissions to this database.}

\item{databaseDescription}{Provide a short description of the database.}

\item{accountName}{Name of the blob container in question}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                            An object containing per-section tables and parameters for the given data source
}\if{html}{\out{</div>}}
}
\description{
\code{ReportSections} defines logic for each of the proposed sections of the CHoRUS Report
}
\details{
\code{ReportSections} the proposed structure can be found here: https://github.com/chorus-ai/chorus-cloud-pm/issues/3#issuecomment-2069945869
}
