% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{getPriorReleaseRds}
\alias{getPriorReleaseRds}
\title{Helper functions for CHoRUS Reports}
\usage{
getPriorReleaseRds(connectionDetails, priorOid, databaseName)
}
\arguments{
\item{connectionDetails}{connectionDetails object for DatabaseConnector}

\item{databaseName}{Name of the data site under review}

\item{prior_oid}{oid generated during upload of RDS object into postgres}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                             An object containing results from the most recent run
}\if{html}{\out{</div>}}
}
\description{
\code{Helper} provides helper functions to create the reports for CHoRUS Data Sites
}
\details{
\code{Helper} provides helper functions to create the reports for CHoRUS Data Sites
}
