% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{list_blobs}
\alias{list_blobs}
\title{Helper functions for CHoRUS Reports}
\usage{
list_blobs(container_url, sas, container_name)
}
\arguments{
\item{container_url}{Web location of Azure storage account}

\item{sas}{SAS key for the Azure storage account}

\item{container_name}{Name of the blob container in question}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                               An object containing the files and their metadata in the relevant storage container on Azure
}\if{html}{\out{</div>}}
}
\description{
\code{Helper} provides helper functions to create the reports for CHoRUS Data Sites
}
\details{
\code{Helper} provides helper functions to create the reports for CHoRUS Data Sites
}
